Rename `:std:package-protected` to ':std:visibility'
