--- std/package-protected-lint/src/main/resources/META-INF/com.android.tools.lint.client.api.IssueRegistry
@@ -1,1 +1,0 @@
-br.com.orcinus.orca.std.packageprotected.lint.PackageProtectedIssueRegistry
